+++ mobile/android/android-components/docs/changelog.md
@@ -12,6 +12,10 @@ permalink: /changelog/
 * [Gecko](https://github.com/mozilla-mobile/android-components/blob/master/buildSrc/src/main/java/Gecko.kt)
 * [Configuration](https://github.com/mozilla-mobile/android-components/blob/master/buildSrc/src/main/java/Config.kt)
 
+* **feature-share**
+  * Added database to store recent apps
+  * Added `RecentAppsStorage` to handle storing and retrieving most-recent apps
+
 # 32.0.0
 
 * [Commits](https://github.com/mozilla-mobile/android-components/compare/v31.0.0...v32.0.0)